requirements.txt
astronomer-cosmos
dbt-postgres
psycopg2-binary
dbt-core